+++ example-config.json
@@ -4,5 +4,7 @@
     "mysqlhost" : "",
     "mysqluser" : "",
     "mysqlpassword" : "",
-    "mysqldatabase" : ""
+    "jwtSecret": "",
+    "mysqldatabase" : "",
+    "refreshTokenSecret": ""
 }
