--- other/hash-decode.js
@@ -1,6 +1,0 @@
-const { hashSync, genSaltSync, compareSync } = require('bcrypt')
-
-function hashPassword(password) {
-    const salts = genSaltSync(10);
-
-}
